implement Identifiable where id getter is already defined
I found 17 definitions in TootSDK that have an id getter. 6 of them implement Identifiable, and I've added Identifiable to the remaining 11. A practical convenience of this is that it allows iteration with SwiftUI ForEach, e.g. I'm currently doing that with Mention and MediaAttachment and adding an empty extension to implement Identifiable.

https://github.com/TootSDK/TootSDK/discussions/108